requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
paho-mqtt==1.6.1
psycopg2-binary==2.9.10
requests==2.32.3
urllib3==2.3.0
